Move tc_array_size and array_range_verify to type_set.rs
It belongs to TypeSet and not directly called in type_infer.rs.
