+++ parser/src/type_infer.rs
@@ -653,45 +653,6 @@ fn forward_lvalue<'src, 'b, 'native>(
     }
 }
 
-pub(crate) fn tc_array_size(value: &ArraySize, target: &ArraySize) -> Result<(), String> {
-    match (value, target) {
-        (_, ArraySize::Any) => {}
-        (ArraySize::Fixed(v_len), ArraySize::Fixed(t_len)) => {
-            if v_len != t_len {
-                return Err(format!(
-                    "Array size is not compatible: {v_len} cannot assign to {t_len}"
-                ));
-            }
-        }
-        (ArraySize::Range(v_range), ArraySize::Range(t_range)) => {
-            array_range_verify(v_range)?;
-            array_range_verify(t_range)?;
-            if t_range.end < v_range.end || v_range.start < t_range.start {
-                return Err(format!(
-                    "Array range is not compatible: {value} cannot assign to {target}"
-                ));
-            }
-        }
-        (ArraySize::Fixed(v_len), ArraySize::Range(t_range)) => {
-            array_range_verify(t_range)?;
-            if *v_len < t_range.start || t_range.end < *v_len {
-                return Err(format!(
-                    "Array range is not compatible: {v_len} cannot assign to {target}"
-                ));
-            }
-        }
-        (ArraySize::Any, ArraySize::Range(t_range)) => {
-            array_range_verify(t_range)?;
-        }
-        _ => {
-            return Err(format!(
-                "Array size constraint is not compatible between {value:?} and {target:?}"
-            ));
-        }
-    }
-    Ok(())
-}
-
 fn tc_coerce_type<'src>(
     value: &TypeSet,
     target: &TypeDecl,
@@ -706,15 +667,6 @@ fn tc_coerce_type<'src>(
         return Err(TypeCheckError::indeterminant_type(span, ctx.source_file));
     }
     Ok(res)
-}
-
-fn array_range_verify(range: &std::ops::Range<usize>) -> Result<(), String> {
-    if range.end < range.start {
-        return Err(format!(
-            "Array size has invalid range: {range:?}; start should be less than end"
-        ));
-    }
-    Ok(())
 }
 
 fn tc_stmt_forward<'src, 'ast, 'native>(
