+++ parser/src/type_set.rs
@@ -1,4 +1,4 @@
-use crate::{interpreter::RetType, type_decl::ArraySize, type_infer::tc_array_size, TypeDecl};
+use crate::{interpreter::RetType, type_decl::ArraySize, TypeDecl};
 
 #[derive(Default, Debug, Clone, PartialEq, Eq)]
 pub enum TypeSet {
@@ -421,4 +421,52 @@ fn array_size_to_string(this: &(Box<TypeSet>, ArraySize)) -> String {
         ArraySize::Range(range) => format!("[{}; {:?}]", this.0, range),
         _ => format!("[{}]", this.0),
     }
+}
+
+fn tc_array_size(value: &ArraySize, target: &ArraySize) -> Result<(), String> {
+    match (value, target) {
+        (_, ArraySize::Any) => {}
+        (ArraySize::Fixed(v_len), ArraySize::Fixed(t_len)) => {
+            if v_len != t_len {
+                return Err(format!(
+                    "Array size is not compatible: {v_len} cannot assign to {t_len}"
+                ));
+            }
+        }
+        (ArraySize::Range(v_range), ArraySize::Range(t_range)) => {
+            array_range_verify(v_range)?;
+            array_range_verify(t_range)?;
+            if t_range.end < v_range.end || v_range.start < t_range.start {
+                return Err(format!(
+                    "Array range is not compatible: {value} cannot assign to {target}"
+                ));
+            }
+        }
+        (ArraySize::Fixed(v_len), ArraySize::Range(t_range)) => {
+            array_range_verify(t_range)?;
+            if *v_len < t_range.start || t_range.end < *v_len {
+                return Err(format!(
+                    "Array range is not compatible: {v_len} cannot assign to {target}"
+                ));
+            }
+        }
+        (ArraySize::Any, ArraySize::Range(t_range)) => {
+            array_range_verify(t_range)?;
+        }
+        _ => {
+            return Err(format!(
+                "Array size constraint is not compatible between {value:?} and {target:?}"
+            ));
+        }
+    }
+    Ok(())
+}
+
+fn array_range_verify(range: &std::ops::Range<usize>) -> Result<(), String> {
+    if range.end < range.start {
+        return Err(format!(
+            "Array size has invalid range: {range:?}; start should be less than end"
+        ));
+    }
+    Ok(())
 }
